move plotting scripts to directory
