+++ dijet/plotting/base.py
