+++ law.cfg
@@ -8,7 +8,8 @@ inherit: $CF_BASE/law.cfg
 
 columnflow.tasks.cms.inference
 columnflow.tasks.cms.external
-dijet.tasks.{alpha,jer,root,plot_asymm}
+dijet.tasks.{alpha,jer,root}
+dijet.plotting.{asymmetry}
 
 
 [logging]
